FIXUP: sched/cpufreq_sched: Clear __sched_energy_freq only when no policy is using sched governor
The current implementation of SchedDVFS is turned off completely if a
cpufreq policy is completely offline. This is because the way the
{set|clear}_sched_energy_freq() functions are implemented only allows
the __sched_energy_freq static_key to be either 1 or 0, while there
are two possible cpufreq policies (one for each cluster) on Oak. The
__sched_energy_freq static_key acts like a switch for the SchedDVFS
hook in the CFS scheduler. So whenever a cluster is turned off, which
sets the corresponding cpufreq policy offline, will result in the
__sched_energy_freq being set to 0.

Signed-off-by: Ricky Liang <jcliang@chromium.org>

BUG=chrome-os-partner:45243
TEST=Build kernel and run on Oak rev3. Disable the big cluster with:

     echo 0 > /sys/devices/system/cpu/cpu2/online
     echo 0 > /sys/devices/system/cpu/cpu3/online

     and run:

     watch -n 0.5 "cat /sys/devices/system/cpu/cpu*/cpufreq/stats/time_in_state"

     to make sure SchedDVFS is still adjusting the LITTLE cores'
     frequency.

Change-Id: Id51ef2138f3f38c4e1b585b6c1fd4bc395ec9ea5
(cherry picked from commit 3a91ef4879eefd3980593039b72fe493dc091cd6)