create-patch: add context below to work with busybox patch
It turns out that GNU patch is happy to only have one line of context
above the inserted area and zero lines below, because of its fuzzy match
feature. However, busybox patch is crippled and terrible, and thus must
have context above and below the insertion point. So, we grudgingly add
it to our manually crafted minimal patch.

Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>